# contracts/openapi/devpath-openapi.yaml
openapi: 3.0.3
info:
  title: DevPath Implemented API
  version: 0.11.0
  description: >
    Implemented contract for authenticated identity, repository synchronization, deterministic analysis,
    rule evaluation, Skill Matrix, career catalog, career readiness, deterministic recommendation,
    learning roadmap, and company catalog slices. Provider access tokens and unimplemented APIs are excluded.
servers:
  - url: http://localhost:8080
    description: Local backend
paths:
...
  /api/v1/repositories/{repositoryId}/sync:
    post:
      x-api-id: API-REP-006
      summary: Queue a durable full repository synchronization
      operationId: synchronizeRepository
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "202":
          description: Existing or newly queued owner-scoped synchronization job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositorySyncJobEnvelope"}
        "400":
          description: Invalid idempotency key or repository lifecycle.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "503":
          description: GitHub permission re-verification is unavailable.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repository-sync-jobs/{jobId}:
    get:
      x-api-id: API-INT-007
      summary: Retrieve an owner-scoped repository synchronization job
      operationId: getRepositorySyncJob
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Current durable job state and safe failure detail.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositorySyncJobEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Job not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/snapshots:
    get:
      x-api-id: API-REP-007
      summary: List immutable snapshots for an owned repository
      operationId: listRepositorySnapshots
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Newest-first immutable snapshot metadata.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositorySnapshotListEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/snapshots/{snapshotId}:
    get:
      x-api-id: API-REP-008
      summary: Retrieve immutable snapshot metadata for an owned repository
      operationId: getRepositorySnapshot
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: snapshotId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Immutable snapshot metadata and deterministic content hash.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositorySnapshotEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Snapshot not found within the owned repository.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/technologies:
    get:
      x-api-id: API-REP-009
      summary: Retrieve normalized technologies from the current owned snapshot
      operationId: getRepositoryTechnologies
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Deterministic language, framework, and database evidence; no official score is calculated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TechnologySummaryEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository or synchronized snapshot not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/evidence:
    get:
      x-api-id: API-REP-011
      summary: Retrieve deterministic engineering evidence from the current owned snapshot
      operationId: getRepositoryEvidence
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Measurable evidence signals without official scores.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryEvidenceSummaryEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository or synchronized snapshot not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
  /api/v1/rule-evaluations/{evaluationId}:
    get:
      x-api-id: API-ANA-004
      summary: Retrieve an immutable deterministic rule evaluation owned by the current user
      operationId: getRuleEvaluation
      tags: [Rule]
      security: [{sessionCookie: []}]
      parameters:
        - {name: evaluationId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Completed deterministic evaluation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RuleEvaluationEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Evaluation not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/rule-evaluations/{evaluationId}/score-breakdown:
    get:
      x-api-id: API-ANA-005
      summary: Retrieve the authoritative category and rule score breakdown
      operationId: getRuleEvaluationScoreBreakdown
      tags: [Rule]
      security: [{sessionCookie: []}]
      parameters:
        - {name: evaluationId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Deterministic score breakdown with calculation trace.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RuleScoreBreakdownEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Evaluation not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/rule-evaluations/{evaluationId}/evidence:
    get:
      x-api-id: API-ANA-006
      summary: Retrieve owner-authorized evidence linked to an immutable evaluation
      operationId: getRuleEvaluationEvidence
      tags: [Rule]
      security: [{sessionCookie: []}]
      parameters:
        - {name: evaluationId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Evidence records and their contributing rules.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RuleEvidenceListEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Evaluation not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DEVPATH_SESSION
      description: Opaque HttpOnly server-session identifier. It is never read by browser JavaScript.
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Server-issued CSRF token. This is not an authentication credential.
      schema:
        type: string
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Owner-scoped stable key for one repository synchronization command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  schemas:
    ApiMetadata:
      type: object
      additionalProperties: false
      required: [requestId, apiVersion, timestamp]
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          example: v1
        timestamp:
          type: string
          format: date-time
...
    RepositorySyncJob:
      type: object
      additionalProperties: false
      required: [jobId, jobType, status, phase, progressPercent, attemptCount, maxAttempts, submittedAt, startedAt, completedAt, pollingUrl, resultResourceUrl, errorCode, errorMessage, retryable]
      properties:
        jobId: {type: string, format: uuid}
        jobType: {type: string, enum: [REPOSITORY_SYNC]}
        status: {type: string, enum: [queued, running, succeeded, failed, cancelled, expired]}
        phase: {type: string, enum: [QUEUED, COLLECTING, RETRY_WAIT, COMPLETED, FAILED]}
        progressPercent: {type: integer, minimum: 0, maximum: 100}
        attemptCount: {type: integer, minimum: 0}
        maxAttempts: {type: integer, minimum: 1}
        submittedAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        pollingUrl: {type: string}
        resultResourceUrl: {type: string, nullable: true}
        errorCode: {type: string, nullable: true}
        errorMessage: {type: string, nullable: true}
        retryable: {type: boolean}
    RepositorySyncJobEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RepositorySyncJob"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    RepositorySnapshot:
      type: object
      additionalProperties: false
      required: [snapshotId, repositoryId, capturedAt, sourceRevision, status, immutable, contentHash, branchCount, commitCount]
      properties:
        snapshotId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        capturedAt: {type: string, format: date-time}
        sourceRevision: {type: string, minLength: 40, maxLength: 64}
        status: {type: string, enum: [READY, FAILED, SUPERSEDED, DELETED_BY_POLICY]}
        immutable: {type: boolean}
        contentHash: {type: string, pattern: "^[a-f0-9]{64}$"}
        branchCount: {type: integer, minimum: 0}
        commitCount: {type: integer, minimum: 0}
    RepositorySnapshotList:
      type: object
      additionalProperties: false
      required: [snapshots]
      properties:
        snapshots:
          type: array
          items: {$ref: "#/components/schemas/RepositorySnapshot"}
    RepositorySnapshotEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RepositorySnapshot"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    RepositorySnapshotListEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RepositorySnapshotList"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    DetectedTechnology:
      type: object
      additionalProperties: false
      required: [name, category, evidenceLabel, byteCount, percentage, taxonomyStatus, evidenceType, evidencePaths]
      properties:
        name: {type: string, minLength: 1, maxLength: 128}
        category: {type: string, enum: [LANGUAGE, FRAMEWORK, DATABASE]}
        evidenceLabel: {type: string, minLength: 1, maxLength: 255}
        byteCount: {type: integer, format: int64, minimum: 0, nullable: true}
        percentage: {type: number, format: double, minimum: 0, maximum: 100, nullable: true}
        taxonomyStatus: {type: string, enum: [SUPPORTED, UNSUPPORTED]}
        evidenceType: {type: string, enum: [PROVIDER_LANGUAGE_STATISTICS, DEPENDENCY_DECLARATION]}
        evidencePaths:
          type: array
          items: {type: string, minLength: 1, maxLength: 500}
    TechnologySummary:
      type: object
      additionalProperties: false
      required: [repositoryId, snapshotId, extractorVersion, taxonomyVersion, primaryLanguage, technologies]
      properties:
        repositoryId: {type: string, format: uuid}
        snapshotId: {type: string, format: uuid}
        extractorVersion: {type: string}
        taxonomyVersion: {type: string}
        primaryLanguage: {type: string, nullable: true}
        technologies:
          type: array
          items: {$ref: "#/components/schemas/DetectedTechnology"}
    TechnologySummaryEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/TechnologySummary"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    EvidenceSignal:
      type: object
      additionalProperties: false
      required: [signalKey, label, present, count, observedValue, evidencePaths]
      properties:
        signalKey: {type: string, pattern: "^[A-Z0-9_]+$"}
        label: {type: string, minLength: 1, maxLength: 128}
        present: {type: boolean}
        count: {type: integer, minimum: 0}
        observedValue: {type: string, nullable: true, maxLength: 255}
        evidencePaths:
          type: array
          maxItems: 20
          items: {type: string, minLength: 1, maxLength: 1000}
    EvidenceCategory:
      type: object
      additionalProperties: false
      required: [category, label, signals]
      properties:
        category: {type: string, enum: [ARCHITECTURE, DATABASE, TESTING, DEVOPS, DOCUMENTATION, ACTIVITY]}
        label: {type: string}
        signals:
          type: array
          items: {$ref: "#/components/schemas/EvidenceSignal"}
    RepositoryEvidenceSummary:
      type: object
      additionalProperties: false
      required: [repositoryId, snapshotId, extractorVersion, categories]
      properties:
        repositoryId: {type: string, format: uuid}
        snapshotId: {type: string, format: uuid}
        extractorVersion: {type: string}
        categories:
          type: array
          items: {$ref: "#/components/schemas/EvidenceCategory"}
    RepositoryEvidenceSummaryEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RepositoryEvidenceSummary"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    RuleResult:
      type: object
      additionalProperties: false
      required: [ruleId, ruleVersion, status, rawValue, score, weight, formulaId, trace, evidenceReferences]
      properties:
        ruleId: {type: string}
        ruleVersion: {type: string}
        status: {type: string, enum: [PASSED, FAILED, PARTIAL, SKIPPED, ERROR]}
        rawValue: {type: number, format: double, minimum: 0}
        score: {type: number, format: double, minimum: 0, maximum: 100}
        weight: {type: number, format: double, minimum: 0, exclusiveMinimum: true, maximum: 1}
        formulaId: {type: string}
        trace: {type: string, maxLength: 1000}
        evidenceReferences:
          type: array
          items: {type: string}
    RuleCategoryScore:
      type: object
      additionalProperties: false
      required: [category, score, weight, confidence, ruleResults, missingEvidence]
      properties:
        category: {type: string, enum: [LANGUAGE, FRAMEWORK, DATABASE, ARCHITECTURE, TESTING, DEVOPS, DOCUMENTATION, ACTIVITY]}
        score: {type: number, format: double, minimum: 0, maximum: 100}
        weight: {type: number, format: double, minimum: 0, exclusiveMinimum: true, maximum: 1}
        confidence: {type: number, format: double, minimum: 0, maximum: 100}
        ruleResults:
          type: array
          items: {$ref: "#/components/schemas/RuleResult"}
        missingEvidence:
          type: array
          items: {type: string}
    RuleEvidenceSummary:
      type: object
      additionalProperties: false
      required: [evidenceCount, rulesWithEvidence, missingEvidenceCount]
      properties:
        evidenceCount: {type: integer, format: int32, minimum: 0}
        rulesWithEvidence: {type: integer, format: int32, minimum: 0}
        missingEvidenceCount: {type: integer, format: int32, minimum: 0}
    RuleEvaluation:
      type: object
      additionalProperties: false
      required: [evaluationId, snapshotId, ruleSetVersionId, ruleSetVersion, formulaLibraryVersion, extractorVersion, overallScore, confidence, evidenceSummary, categoryScores, warnings, completedAt]
      properties:
        evaluationId: {type: string, format: uuid}
        snapshotId: {type: string, format: uuid}
        ruleSetVersionId: {type: string, format: uuid}
        ruleSetVersion: {type: string}
        formulaLibraryVersion: {type: string}
        extractorVersion: {type: string}
        overallScore: {type: number, format: double, minimum: 0, maximum: 100}
        confidence: {type: number, format: double, minimum: 0, maximum: 100}
        evidenceSummary: {$ref: "#/components/schemas/RuleEvidenceSummary"}
        categoryScores:
          type: array
          items: {$ref: "#/components/schemas/RuleCategoryScore"}
        warnings:
          type: array
          items: {type: string}
        completedAt: {type: string, format: date-time}
    RuleEvaluationEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RuleEvaluation"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    RuleScoreBreakdown:
      type: object
      additionalProperties: false
      required: [evaluationId, overallScore, confidence, categories]
      properties:
        evaluationId: {type: string, format: uuid}
        overallScore: {type: number, format: double, minimum: 0, maximum: 100}
        confidence: {type: number, format: double, minimum: 0, maximum: 100}
        categories:
          type: array
          items: {$ref: "#/components/schemas/RuleCategoryScore"}
    RuleScoreBreakdownEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RuleScoreBreakdown"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    RuleEvidence:
      type: object
      additionalProperties: false
      required: [evidenceId, ruleId, contributionRole, evidenceType, sourceReference, observedFactSummary, confidence]
      properties:
        evidenceId: {type: string, format: uuid}
        ruleId: {type: string}
        contributionRole: {type: string, enum: [DIRECT, SUPPORTING, MISSING]}
        evidenceType: {type: string, enum: [REPOSITORY_PATH, LANGUAGE_STATISTIC, SNAPSHOT_SIGNAL]}
        sourceReference: {type: string, maxLength: 1200}
        observedFactSummary: {type: string, maxLength: 500}
        confidence: {type: number, format: double, minimum: 0, maximum: 100}
    RuleEvidenceList:
      type: object
      additionalProperties: false
      required: [evaluationId, evidence]
      properties:
        evaluationId: {type: string, format: uuid}
        evidence:
          type: array
          items: {$ref: "#/components/schemas/RuleEvidence"}
    RuleEvidenceListEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RuleEvidenceList"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, metadata]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
        metadata:
          $ref: "#/components/schemas/ApiMetadata"